Add fix so jQuery doesn't conflict with other Javascript libraries.
I think I finally got this one working.  The saving of the temporary variable has to happen *before* jQuery loads, or else jQuery has already stomped on the '$' variable that another library might have used.  This patch moves the temporary saving of this variable the HTML template so it tucks it aside before jQuery loads and puts it back after we're done.  Tested and works with jQuery, mootools, and prototype.