Don't blow up if INSTALLED_APPS is a tuple
Since r17158 in Django trunk, INSTALLED_APPS is now not always co-erced to
a list. In my projects I prefer to have INSTALLED_APPS as a tuple as it
reinforces that it is not modifiable at runtime. However, this blows up
in the VersionDebugPanel as it currently assumes it is a list.

Signed-off-by: Chris Lamb <lamby@debian.org>