Make it possible to use remoteCall via POST
By using remoteCall CSS class on the `a` tag in the panel template
it is possible to instruct DDT to load content of the HTTP request
into the new panel and put it on the front of the page.

This change allows to configure this call to be a POST request instead
of a GET, which may be desirable in case of long params.

Parameters that need to be sent should be rendered as tag
attributes of `a.remoteCall` prefixed with "data-", i.e.:

    <a class="remoteCall remoteCallPost" href="..." data-var="1">...</a>

will result in sending var=1 in the POST body.