Exempt Ajax panel cache operations from tracking by Cache panel
Cache panel displays arguments that were sent to the cache methods
(they are rendered in the toolbar's template). And because Ajax
panel stores toolbar's templates in the cache (by calling cache.set)
this leads to a geometric grow in the amount of data being stored
as each subsequent template sent to the cache contains content of all the
templates stored previously (as rendered by the Cache panel).

This can quickly lead to template sizes in tens of megabytes after
only a handful of AJAX requests.

On top of that, that data isn't actually interesting from the debugging
point of view.