Don't render list of requests for nothing
The list of requests that should be displayed in the Ajax
panel changes dynamically with each request. Toolbar's DOM
trees for the old requests shouldn't hold this list as it
may be obsolete by the time such a tree is loaded into the
browser (e.g. when going back to examine old request) and
up to date list is always fetched via Ajax call anyway.