Put debug_toolbar's url patterns first
Some applications require using a catch-all urlpattern that must come
last in the list of urlpatterns. If the __debug__ urls come after it,
they'll never be hit.

Putting the debug_toolbar urls first fixes this, but could possibly
shadow an existing url, but the old version of debug toolbar did the
same thing.