Propagate mouseup events when not dragging toolbar
In Chrome (this does not affect Firefox) onchange events on multi-selects are not fired all of the time e.g. with the following template:

<body>
    <select multiple="multiple" onchange="console.log('clicked');">
        <option  value="1">1</option>
    </select>
</body>

Ctrl click on the option and view the js log. many of the events will not register. Therefore stop propagation and prevent default action only when handle is dragged.