Work around spurious mousemove on Chrome.
When clicking on an element in Chrome, spurious mousemove events can be
generated when the cursor hasn't actually moved.  This can prevent the
toolbar from being shown when the handle is clicked because the code
sees the mousemove event and thinks a handle drag is happening.  Work
around this by only running the handle drag code if the cursor has moved
since the mousedown event.