toolbar.js: Ensure toolbar handle cannot be dragged outside window
Previously the toolbar handle could be dragged beyond the top and bottom
of the window where it could get lost.

If the handle position outside the window border was saved, the only way
to get the handle back that I found would be to open the inspector and
manually change the handle's offset to make it visible once again.

This change restricts the handle so that it cannot be dragged above the
top or below the bottom of the window area.

Using `event.clientY` in order to get the cursor position relative to
the viewport instead of the document (as is the case with
`event.pageY`). This allows the handle to be dragged when the page is
scrolled.

Closes #634.