Support disabling panels with the names in the docs.
The docs hide the fact that the SQLPanel and TemplatesPanel live in
a panel module. The DISABLE_PANELS config option previously required
that panel module to be included in the path in order for the panel
to be disabled.

This change will check the panel's path against that config set and
then check it again with the .panel. replaced as just . against the
config set.

Doing this replace blindly will support moving any other panel from
it's own module into a package like the SQLPanel and TemplatesPanel.

Fixes #687.