Create generate_stats, an optional method similar to process_response.
generate_stats is not always called. It will only be called when
the toolbar is going to insert itself into the response. The method
process_response will always be called and should be used for panels
that need to interrupt the request such as the RedirectsPanel.

This update is to address issue #517.