Ignore the call to the profilers disable method in the ProfilingPanel.
While this hasn't caused any issues with the actual usage of the
toolbar, it has caused an unexpected error within the unit tests.
This is because the dictionary returned by stats is converted into
a list. The order of this list isn't consistent, while the logic
expects it to which gave it the random feel to it.

Ignoring the call to the disable should cause the view function
to be the only function with zero callers.