Improve documentation around generate_stats.
Added a line for the 1.4 release which will contain this new
method.
Revised the sphinx method references for generate_stats and
process_response to actually link to each other. Following
the existing format to link to process_response lead to a
link being generated to Django's actual process_response
method on the middleware. I had to specifically mention the
toolbar's process_response and then updated the generate_stats
reference to be consistent between the two methods.