Make the fix for #591 even more focused
When we undefine the entire window.define, modules loaded asynchronously
can randomly fail to register themselves. However, jQuery 2+ checks for
define.amd before registering itself as an AMD module, not just define.
So, by unsetting window.define.amd instead of window.define, we don't
break define() for other modules and we still prevent the page-global
jQuery from registering itself as an AMD module, which was the ultimate
goal of #591.