Fix missing query-actions buttons for queries with date/datetime params.
The param `strings_only=True` caused params with type `time`, `date` and
`datetime` to not be converted to string by `NormalCursorWrapper._decode()`,
and in turn causes the json dumping to fail.

This change keeps `strings_only=True` for every type but datetime, date and
time, enabling conversion to string in those cases.

Fixes #882