support structured types when recording SQL
If Django's support for array/JSON/HStore Postgres fields is used then
the SQL query can no-longer be explained because array and dictionary
parameters are forced into textual representations.

Teach the _decode method of panels.sql.tracking.NormalCursorWrapper how
to recurse into list, tuple and dictionary objects. It's plausible that
more sequence/mapping types may need to be added in future or one could
use the abstract base classes from [1] but then a special case would
need to be made for all string-like objects; strings are sequences but
should not be recursed as other sequences.

[1] https://docs.python.org/3/library/collections.abc.html

Closes #796. Closes #1001.