Restructure the Panel class to execute more like the MIDDLEWARE
This allows the ProfilingPanel to no longer skip remaining panels or
middlewares.

This change is backwards incompatible. Third party panels will need to
adjust to use this new architecture. It removes support for Django
1.11's deprecated MIDDLEWARE_CLASSES setting.

The Panel.__init__() method is now passed a get_response argument. The
Panel.process_request() method must now always return a response.
Usually this is the response returned by get_response() but may also
return a different response as is the case in the ``RedirectsPanel``.

Panel.process_response() and Panel.process_view() have been removed.

Fixes #1135