Run eslint during CI
ESLint is a tool for linting JavaScript code, with the goal of making
code more consistent and avoiding bugs. In many ways, it is similar to
flake8, but for JavaScript.

This replaces the previous use of JSHint, which was not checked during
CI. ESLint is the more popular choice in the JavaScript community and
offers more rules and features.