Explicitly set the 'box-sizing' property of the .djDebugClose to 'content-box'.
It is common for some CSS and reset frameworks to use set the 'box-sizing' to 'border-box' to all elements in the page. This causes the close button inside each panel to show up in the wrong way.

This set the 'box-sizing' attribute of the .djDebugClose class to always be  'content-box'.