Automate screenshot generation in README.rst
The current screenshot is outdated. It is running on an older version of Django
Debug Toolbar and an end-of-life Django 1.6. Some newer panels, such as the
history panel, are missing.

The new script example/screenshot.py uses Selenium to generate a screenshot
using the latest version of Django and Django Debug Toolbar. Unlike the
previous screenshot which displayed macOS window decorations, this one is
OS-agnostic. The script generates the image with an exact resolution, so
README.rst no longer needs to re-specify dimensions.

The generated image is optimized using the optipng command line utility.

The contributing docs have been updated to include running this script when
making a new release so the screenshot will continue to remain up to date.