Simplify fetch() calls in JavaScript
The fetch() init object does not have a property "url", so remove it.

The element being either a button or a element is mutually exclusive, so
use an "else if".

When a button is pressed, an enclosing form element already exist, so
remove the guard.

Assign the object properties in the object literal instead of after.