Change .djDebugClose <a> tag to more semantic <button type="button">
Anchor tags are for hyperlinks. The close button does not link to (nor
contains) a new URL. The href was previously specified as an empty
string.

Replace these with actual buttons. The buttons continue to be styled the
same as before. The CSS was simplified a bit due to the change in the
element.