Display all fetch() errors in ajax()
Previously, only successful requests with non-200 responses would
display as an error. But fetch() can fail for other reasons too. For
example, if the network connection is lost, fetch() fails with:

    TypeError: NetworkError when attempting to fetch resource.

This is now displayed rather than showing the spinner indefinitely.

The error is re-thrown in JavaScript so it displays in the JavaScript
console and can interact with the browser's developer tools.