Simplify CSS classes and rules on th/td elements
These CSS classes don't style th element, but td elements. So remove
from th elements for slightly smaller HTML.

The rule:

    #djDebug .djdt-panelContent thead th {
        white-space: nowrap;
    }

duplicates the rule on line 350 with the same selector. A it is
redundant, remove it.

djdt-query had no CSS and no JavaScript associated with it. As it has no
effect, remove it.

djdt-timeline only adds a width. There already exist a series of generic
classes to set widths "djdt-width-*', so use this pattern instead of a
new one-off CSS class.