Fixed issue with toolbar expecting urls to start with __debug__.
When the history panel was added and we expanded the number of
urls we were instrumenting, there was a check added to exclude
the requests to the toolbar's views. This was done by checking
the path of the request to see if it started with __debug__.
However, the documentation states that is not a requirement.
This change checks the urls to see if the namespace of the
resolved url matches the toolbar's namespace.

This still isn't perfect as another url path could contain the
last namespace of djdt, but this is much less likely than a user
defining the toolbar's urls at a different path than /__debug__/