Merge pull request from GHSA-pghf-347x-c2gj
* Fix CVE-2021-30459 by creating signature from all data fields.

Create a signature based on all fields in the form and attach
to validate that the data being sent back is what the server
generated initially.

Change the hashing algorithm to SHA256

Force the values to a string for signing.

Remove hashing mechanism from forms.

Support sha1 algorithm for django < 3.1

* Bump version to 3.2.1