Use only pre-commit for style linters and fixers (#1535)
* Use only pre-commit for style linters and fixers

The method of using the `Makefile`, running under `tox`, duplicated the work of pre-commit and pre-commit.ci. Additionally tox did not use the same pinned versions and arguments as in `.pre-commit-config.yaml`, so there is the potential for a version clash to cause the two paths to conflict, e.g. when a new version of Black is released.

Sticking to only `pre-commit` seems sensible, saves some CI time, and allows some files to be removed.

* spelling