Optimize render_stacktrace()
This function is called by the SQL panel to render each frame of each stack trace.

Previously it used a Django template, which is slow. By generating the HTML with pure Python code, we can avoid much of this overhead.

I tried this change on a demo app I built that has 96 total queries, due to N+1 queries. The queries execute in 13ms due to use of SQLite, so the whole view is a negligible concern. Nearly all the view runtime is the toolbar itself. Without this change, the runtime is ~1300ms; with the change it’s ~1100ms. That's a saving of **15%**.

I also checked the appearance of the generated HTML hasn’t changed.