Make .assertValidHTML() parse passed content
Prior to this commit, the BaseTestCase.assertValidHTML() method was not
parsing the content from the provided argument, but was instead parsing
self.panel.content.  For most usages, the passed in content was the same
as self.panel.content, but for the
TemplatesPanelTestCase.test_template_repr() test case it was not,
resulting in an invalid test.