Fix and improve .assertValidHTML() test method (#1597)
* Make .assertValidHTML() parse passed content

Prior to this commit, the BaseTestCase.assertValidHTML() method was not
parsing the content from the provided argument, but was instead parsing
self.panel.content.  For most usages, the passed in content was the same
as self.panel.content, but for the
TemplatesPanelTestCase.test_template_repr() test case it was not,
resulting in an invalid test.

* Fix test failure exposed by previous commit

* Remove unused .assertValidHTML() msg argument

Allows removal of a call to the internal TestCase._formatMessage()
method.

* Tweak .assertValidHTML() exception message

Since the passed in HTML does not necessarily represent the panel
content, change the message to no longer reference "Content".