Cache getframeinfo() results for get_stack()
getframeinfo() is quite expensive as it hits the filesystem multiple
times.  Wrap with a custom least-recently-used cache to reduce this
cost.  It is not possible to use functools.lru_cache directly as using
frame objects as cache keys would cause them (and their associated
locals) to stay in memory indefinitely.  The custom LRU cache takes a
function which transforms the frame argument into a suitable cache key.