The path may not always be a true path for stacktraces.
Occassionally we will get a stacktrace that's an importlib instance string
representation. While we may be able to put the python path or something else
it's likely easier (and more logical) to simply pass that onto the user.

I was unable to reproduce the issue in our tests, so I've mocked the case
in test_importlib_path_issue_1612.

Fixes #1612