Rename SQLPanel context var to control SQL access.
The current build started failing for py3.8+ because the inner
exception SQLQueryTriggered was raised in the test
test_cursor_wrapper_asyncio_ctx.

This was identified as being caused by sql.tracking.recording
being set to false. The name recording made it seem as if
this context var was controlling whether recording was occurring
or not. However, it's true functionality was preventing SQL
queries from being sent during the template panel's processing.

Renaming that context var to allow_sql is more indicative of
its purpose and makes the fixed test change clearer.