Allow integration tests to access the toolbar
Add a mechanism for integration tests to access the toolbar used for a
particular request:  Update DebugToolbar to emit a signal on creation
referencing itself.  Then create and use a custom Django test client
that connects to this signal to capture the toolbar that was created
while the request was being processed, and to store the toolbar on the
response for use by tests.