Replace sqlparse.filters.SerializerUnicode() usage
sqlparse's SerializerUnicode filter does a bunch of fancy whitespace
processing which isn't needed because the resulting string will just be
inserted into HTML.  Replace with a simple EscapedStringSerializer that
does nothing but convert the Statement to a properly-escaped string.

In the process stop the escaping within BoldKeywordFilter to have a
cleaner separation of concerns:  BoldKeywordFilter now only handles
marking up keywords as bold, while escaping is explicitly handled by the
EscapedStringSerializer.