Format frame locals when captured
Before this commit, if ENABLE_STACKTRACES_LOCALS was True, each stack
frame's locals dict would be stored directly in the captured stack
trace, and only converted to a string once the stack trace was rendered.
This means that for panels, such as the SQL Panel, which do not render
the stack traces until later after they have all been captured, it is
quite possible that the values in the locals dicts at the point when the
stack trace was rendered would no longer be the same as they were at the
time when the stack trace was captured, resulting in incorrect values
being displayed to the user.

Fix by converting the locals dict to a string immediately when it is
captured (in the get_stack_trace() function as well as in the deprecated
tidy_stacktrace() function) instead of in the render_stacktrace()
function.