Use Python's built-in html.escape()
Because the token values escaped by BoldKeywordFilter are simply
intermediate values and are not directly included in HTML templates,
use Python's html.escape() instead of django.utils.html.escape() to
eliminate the overhead of converting the token values to SafeString.

Also pass quote=False when calling escape() since the token values will
not be used in quoted attributes.