Don't try to undo cache method monkey patching
Trying to undo the monkey patch of cache methods in the
CachePanel.disable_instrumentation() method is fragile in the presence
of other code which may also monkey patch the same methods (such as
Sentry's Django integration), and there are theoretically situations
where it is actually impossible to do correctly.  Thus once a cache has
been monkey-patched, leave it that way, and instead rely on checking in
the patched methods to see if recording needs to happen.  This is done
via a _djdt_panel attribute which is set to the current panel in the
enable_instrumentation() method and then set to None in the
disable_instrumentation() method.