If wsgi.multiprocess isn't set, render panels on each request.
The likely cause of this is that the application is using ASGI since
wsgi.multiprocess is a required key for a WSGI application. Since the
toolbar currently doesn't support async applications, it's pretty likely
that it won't work for this request.

If you're a developer reading this and you think I'm wrong, you can set
the RENDER_PANELS setting to forcibly control this setting.