Allow more control over tox Selenium tests
Instead of unconditionally enabling Selenium tests for the
py311-dj42-postgresql tox environment, enable them by default for that
environment but allow them to be disabled if the user's
DJANGO_SELENIUM_TESTS environment variable is empty.  This will allow
the user to run

    DJANGO_SELENIUM_TESTS= tox

to run the full tox test suite with Selenium tests disabled.