Make toolbar compatible with `FORCE_SCRIPT_NAME`
Previously, if a project used the `FORCE_SCRIPT_NAME` setting (common when
hosting a Django application on a subdirectory path via a reverse proxy),
the `django.urls.resolve()` call would always raise `Resolver404` in the
middleware. As a result, `is_toolbar_request()` always returned False.

This caused internal toolbar URLs to be inspected, and also indirectly
led to a request loop when refreshing the history panel.

In most cases (if `FORCE_SCRIPT_NAME` is unset), `get_script_prefix()` will
return "/" and the `replace()` will be a no-op.